requirements.txt
amqp==5.2.0
asgiref==3.8.1
asttokens==2.4.1
async-timeout==4.0.3
attrs==24.2.0
Automat==22.10.0
beautifulsoup4==4.12.3
billiard==4.2.0
celery==5.4.0
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
comm==0.2.2
constantly==23.10.4
cron-descriptor==1.4.3
cryptography==43.0.0
cssselect==1.2.0
debugpy==1.8.5
decorator==5.1.1
defusedxml==0.7.1
Django==5.0.8
django-celery-beat==2.6.0
django-timezone-field==7.0
djangorestframework==3.15.2
exceptiongroup==1.2.2
executing==2.0.1
filelock==3.15.4
flower==2.0.1
gunicorn==23.0.0
humanize==4.10.0
hyperlink==21.0.0
idna==3.7
incremental==24.7.2
ipykernel==6.29.5
ipython==8.26.0
itemadapter==0.9.0
itemloaders==1.3.1
jedi==0.19.1
jmespath==1.0.1
jupyter_client==8.6.2
jupyter_core==5.7.2
kombu==5.4.0
lxml==5.3.0
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
packaging==24.1
parsel==1.9.1
parso==0.8.4
pexpect==4.9.0
platformdirs==4.2.2
prometheus_client==0.20.0
prompt_toolkit==3.0.47
Protego==0.3.1
psutil==6.0.0
psycopg==3.2.1
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyDispatcher==2.0.7
Pygments==2.18.0
PyMySQL==1.1.1
pyOpenSSL==24.2.1
PySocks==1.7.1
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
pyzmq==26.1.0
queuelib==1.7.0
redis==5.0.8
requests==2.32.3
requests-file==2.1.0
Scrapy==2.11.2
service-identity==24.1.0
six==1.16.0
soupsieve==2.6
sqlparse==0.5.1
stack-data==0.6.3
tldextract==5.1.2
tomli==2.0.1
tornado==6.4.1
traitlets==5.14.3
Twisted==24.7.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
vine==5.1.0
w3lib==2.2.1
wcwidth==0.2.13
zope.interface==7.0.1
